% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPM.R
\name{as.MGPM}
\alias{as.MGPM}
\title{If necessary, convert an object to an MGPM object.
This is an S3 generic function.}
\usage{
as.MGPM(o, env = attr(o, "env"), treeEDExpr = attr(o, "treeEDExpr"))
}
\arguments{
\item{o}{an object.}

\item{env}{a \code{\link{MGPMEnvironment}} object. Default: \code{attr(o, "env")}.}

\item{treeEDExpr}{a character string denoting a ED-expression.
Default: \code{attr(o, "treeEDExpr")}.
\code{\link{PCMBase::PCMTreeEvalNestedEDxOnTree()}}.}
}
\value{
an MGPM object corresponding to \code{o}. If \code{o} is already
a \code{\link{MGPM}} object, it is returned as is. Otherwise, a
\code{\link{MGPM}} object is constructed from \code{o} and,
optionally \code{env} and \code{treeEDExpr}.
}
\description{
If necessary, convert an object to an MGPM object.
This is an S3 generic function.
}
\seealso{
\code{\link{MGPM}}
}
