% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ComposeMixedGaussianFromFits}
\alias{ComposeMixedGaussianFromFits}
\title{Compose a MixedGaussian model}
\usage{
ComposeMixedGaussianFromFits(tree, startingNodesRegimes, modelTypes, k, R,
  mapping, argsMixedGaussian, tableFits, modelTypesInTableFits,
  tableAncestors = NULL, verbose = FALSE)
}
\arguments{
\item{tree, startingNodesRegimes, modelTypes, k, R, mapping, argsMixedGaussian, tableFits, modelTypesInTableFits, tableAncestors, verbose}{Internal use.}
}
\value{
a PCM object.
}
\description{
Compose a MixedGaussian model
}
\details{
This is an internal function that needs to be exported solely for
technical reasons. Not intended to be called by the end user.
}
\keyword{internal}
