% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitRegimesAndModelMappings.R
\name{ComposeMRGFromFits}
\alias{ComposeMRGFromFits}
\title{Compose an MRG model}
\usage{
ComposeMRGFromFits(tree, startingNodesRegimes, modelTypes, k, R, mapping,
  argsMRG, tableFits, tableAncestors = NULL)
}
\description{
Compose an MRG model
}
