% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappingsToPartitions.R
\docType{data}
\name{.cacheEnv}
\alias{.cacheEnv}
\title{Internally used cache environment.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.cacheEnv
}
\description{
Internally used cache environment.
}
\keyword{datasets}
