% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableFits.R
\name{HashCodes}
\alias{HashCodes}
\title{Generate hash codes.}
\usage{
HashCodes(tree, modelTypes, startingNodesRegimesLabels, modelMapping)
}
\arguments{
\item{tree}{a phylo object}

\item{modelTypes}{character vector}

\item{startingNodesRegimesLabels}{a character vector denoting the
partition nodes.}

\item{modelMapping}{an integer or character vector to be matched against
modelTypes}
}
\value{
a character vector.
}
\description{
Generate hash codes.
}
