% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{imapping}
\alias{imapping}
\title{Iterator over combinations with repetions of a given set of models}
\usage{
imapping(mapping, models)
}
\arguments{
\item{mapping}{a vector of elements from models giving the initial combination}

\item{models}{a vector of unique elements to choose from when building the
combinations.}
}
\value{
an iterator object with S3 class c("imapping", "abstractiter", "iter").
Calling repeatedly nextElem on this object iterates over all possible combinations
with repetitions.
}
\description{
Iterator over combinations with repetions of a given set of models
}
\examples{
it <- imapping(c("BM3", "BM3"), c("BM3", "OU3", "JOU3"))
nextElem(it)
nextElem(it)
}
