% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMEnvironment.R
\name{is.MGPMEnvironment}
\alias{is.MGPMEnvironment}
\title{Check if an object is of S3 class 'MGPMEnvironment'}
\usage{
is.MGPMEnvironment(o)
}
\arguments{
\item{o}{an object.}
}
\value{
a logical.
}
\description{
Check if an object is of S3 class 'MGPMEnvironment'
}
