% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMTable.R
\name{is.MGPMTable}
\alias{is.MGPMTable}
\title{Check if an object is a MGPMTable.}
\usage{
is.MGPMTable(o)
}
\arguments{
\item{o}{an R object.}
}
\value{
a logical
}
\description{
Check if an object is a MGPMTable.
}
