% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitMixed.R
\name{is.PCMFitModelMappings}
\alias{is.PCMFitModelMappings}
\title{Check if an object is a PCMFitModelMappings.}
\usage{
is.PCMFitModelMappings(object)
}
\arguments{
\item{object}{an R object.}
}
\value{
logical.
}
\description{
Check if an object is a PCMFitModelMappings.
}
