% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JitterModelParams.R
\name{JitterModelParams}
\alias{JitterModelParams}
\title{Add random white noise to the parameters of a model.}
\usage{
JitterModelParams(model, argsPCMParamLowerLimit = NULL,
  argsPCMParamUpperLimit = NULL, numJitterRootRegimeFit = 1000,
  sdJitterRootRegimeFit = 0.5, numJitterAllRegimeFits = 1000,
  sdJitterAllRegimeFits = 0.5, returnWithinBoundsOnly = TRUE,
  verbose = FALSE)
}
\arguments{
\item{model}{a PCM object.}

\item{argsPCMParamLowerLimit, argsPCMParamUpperLimit}{named lists with
arguments passed to the functions \code{\link{PCMParamLowerLimit}} and
\code{\link{PCMParamUpperLimit}}, respectively. Default: \code{NULL}.}

\item{numJitterRootRegimeFit, numJitterAllRegimeFits}{integers specifying how
many initial parameter vectors should be generated by 'jittering' the
parameters of the root regime of a previously found fit
(such as a clade-fit), and by jittering all regime parameters.
Default: \code{100}.}

\item{sdJitterRootRegimeFit, sdJitterAllRegimeFits}{white noise standard
deviation. Default: \code{0.5}.}

\item{returnWithinBoundsOnly}{logical indicating if only vectors within the
boundaries specified by \code{\link{PCMParamLowerLimit}(model)} and
\code{\link{PCMParamUpperLimit}(model)} should be returned.}

\item{verbose}{logical indicating if debug messages should be printed to the
console. Default: \code{FALSE}.}
}
\value{
a numerical matrix, each row corresponding to a jittered parameter
vector.
}
\description{
Add random white noise to the parameters of a model.
}
