% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitRegimesAndModelMappings.R
\name{MatchModelMapping}
\alias{MatchModelMapping}
\title{Match a model mapping vector against a vector of model types}
\usage{
MatchModelMapping(modelMapping, modelTypes)
}
\arguments{
\item{modelMapping}{a character or integer vector}

\item{modelTypes}{a vector}
}
\value{
an integer vector or stops with an error
}
\description{
Match a model mapping vector against a vector of model types
}
