% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPMTable.R
\name{MGPMTable}
\alias{MGPMTable}
\title{A \code{\link{data.table}} of \code{\link{MGPM}} objects}
\usage{
MGPMTable(o, objectName = NULL, env = NULL, treeEDExpr = "tree",
  verbose = FALSE)
}
\arguments{
\item{o}{an object, such as a character string denoting a path to a '.RData',
'.rda' file where an object named as indicated by \code{objectName} is stored,
or a \code{\link{MGPMEnvironment}} object, or a \code{\link{MGPM}}, or a list of
\code{MGPM} objects or a list of \code{double} vectors (see argument \code{env}).}

\item{objectName}{\code{NULL} or a character string denoting an R object stored
in the file indicated by \code{o}. This argument is ignored if \code{o} is not a
character string (file-path).}

\item{env}{a \code{\link{MGPMEnvironment}} object. This must be specified only if
\code{o} is a list of \code{double} vectors. Otherwise this argument is ignored.}

\item{treeEDExpr}{a character string denoting a ED-expression
(see \code{\link{PCMBase::PCMTreeEvalNestedEDxOnTree}}). Like the \code{env}
argument, this must be specified only if \code{o} is a list of \code{double}
vectors. Otherwise this argument is ignored.}

\item{verbose}{logical (default \code{FALSE}) indicating if some log-info
should be printed to the console.}
}
\value{
a \code{MGPMTable} object.
}
\description{
A \code{\link{data.table}} of \code{\link{MGPM}} objects
}
\details{
This is a S3 generic function. See \code{methods(MGPMTable)} for
implementing methods.
}
