% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitDemoObjects.R
\docType{data}
\name{PCMFitDemoObjects}
\alias{PCMFitDemoObjects}
\title{Demo objects for the PCMFit package}
\format{A list containing simulated trees, trait-values model objects and inference
results.}
\usage{
PCMFitDemoObjects
}
\description{
Demo objects for the PCMFit package
}
\keyword{datasets}
