% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{PCMFitModelMappings}
\alias{PCMFitModelMappings}
\title{Fit regime-assignments to (sub-)trees in a tree with different assigned model types to each regime.}
\usage{
PCMFitModelMappings(X, tree, modelTypes = c("BM__noX0__noSigmae_x",
  "OU__noX0__noSigmae_x"), metaIFun = PCMInfo, positiveValueGuard = Inf,
  lik = NULL, prior = NULL, input.data = NULL, config = NULL,
  fitMappingsPrev = NULL, tableFitsPrev = fitMappingsPrev$tableFits,
  skipFitWhenFoundInTableFits = TRUE, prefixFiles = "fits_",
  maxCladePartitionLevel = 1, maxNumNodesPerCladePartition = Inf,
  minCladeSizes = 25, listPCMOptions = PCMOptions(), argsMRG = NULL,
  argsPCMLowerBound = NULL, argsPCMUpperBound = NULL,
  argsPCMSetOrGetVecParams = NULL, argsConfigOptimAndMCMC1 = NULL,
  argsConfigOptimAndMCMC2 = NULL, numJitterRootRegimeFit = 100,
  sdJitterRootRegimeFit = 0.5, numJitterAllRegimeFits = 100,
  sdJitterAllRegimeFits = 0.5, printFitVectorsToConsole = FALSE,
  setAttributes = TRUE, doParallel = FALSE, verbose = TRUE,
  verbosePCMFit = FALSE, verboseComposeMRGFromFits = FALSE,
  verboseAdaptArgsConfigOptimAndMCMC = FALSE)
}
\value{
an S3 object of class PCMFitModelMappings.
}
\description{
This function performs multiple model fits of mixed regime models
(MRG) mapping different model-types (e.g. BM and OU) to different regimes in a tree and
testing different regime assignments to the branches in the tree (TODO describe algorithm).
}
