% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappingsToCladePartitions.R
\name{PCMFitModelMappingsToCladePartitions}
\alias{PCMFitModelMappingsToCladePartitions}
\title{Fit possible model mappings from a set of model types to a list of clade partitions of a tree}
\usage{
PCMFitModelMappingsToCladePartitions(X, tree, modelTypes, SE = matrix(0,
  nrow(X), PCMTreeNumTips(tree)), listCladePartitions = NULL,
  listAllowedModelTypesIndices = NULL, generatePCMModelsFun = NULL,
  metaIFun = PCMInfo, positiveValueGuard = Inf, lik = NULL,
  prior = NULL, input.data = NULL, config = NULL,
  skipFitWhenFoundInTableFits = TRUE, prefixFiles = "fits_",
  listPCMOptions = PCMOptions(), argsMixedGaussian = NULL,
  argsPCMParamLowerLimit = NULL, argsPCMParamUpperLimit = NULL,
  argsPCMParamLoadOrStore = NULL, argsConfigOptimAndMCMC = NULL,
  numJitterRootRegimeFit = 100, sdJitterRootRegimeFit = 0.5,
  numJitterAllRegimeFits = 100, sdJitterAllRegimeFits = 0.5,
  printFitVectorsToConsole = FALSE, doParallel = FALSE,
  verbose = TRUE, verbosePCMFit = FALSE,
  verboseComposeMixedGaussianFromFits = FALSE,
  verboseAdaptArgsConfigOptimAndMCMC = FALSE)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} and \code{NaN} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Missing values can be either not-available (\code{NA}) or not existing (\code{NaN}).
Thse two values have are treated differently when calculating
likelihoods: see \code{\link{PCMPresentCoordinates}}.}

\item{tree}{a phylo object with N tips.}

\item{SE}{a k x N matrix specifying the standard error for each measurement in
X. Alternatively, a k x k x N cube specifying an upper triangular k x k
Choleski factor of the variance covariance matrix for the measurement error
for each node i=1, ..., N.
Default: \code{matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree))}.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}
}
\description{
Fit possible model mappings from a set of model types to a list of clade partitions of a tree
}
