% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{PCMIteratorMapping2}
\alias{PCMIteratorMapping2}
\title{Iterator over combinations with repetions of a given set of modelTypes}
\usage{
PCMIteratorMapping2(mapping, modelTypes, allowedModelTypesIndices)
}
\arguments{
\item{mapping}{a vector of elements from modelTypes giving the initial combination}

\item{modelTypes}{a vector of unique elements to choose from when building the
combinations.}

\item{allowedModelTypesIndices}{a list of the same length as \code{mapping} with integer vector elements
or NAs. When an element of this list is an integer vector its elements denote unique positions in
modelTypes, i.e. the allowed model-types for the regime at that position in mapping.}
}
\value{
an iterator object with S3 class c("imapping", "abstractiter", "iter").
Calling repeatedly nextElem on this object iterates over all possible combinations
with repetitions of the same length as the argument \code{mapping}.
}
\description{
Iterator over combinations with repetions of a given set of modelTypes
}
\examples{
it <- PCMIteratorMapping(c("BM", "BM"), c("BM", "OU", "JOU"))
iterators::nextElem(it)
iterators::nextElem(it)
}
