% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{PCMLoadMixedGaussianFromFitVector}
\alias{PCMLoadMixedGaussianFromFitVector}
\title{Load a mixed-regime Gaussian model from a fit vector}
\usage{
PCMLoadMixedGaussianFromFitVector(fitVector, modelTypes, k,
  remapModelTypeIndicesInFitVector = seq_along(modelTypes), ...)
}
\description{
Load a mixed-regime Gaussian model from a fit vector
}
