% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{PCMLoadMRGFromFitVector}
\alias{PCMLoadMRGFromFitVector}
\title{Load a mixed-regime Gaussian model from a fit vector}
\usage{
PCMLoadMRGFromFitVector(fitVector, modelTypes, k, ...)
}
\arguments{
\item{fitVector}{a numeric vector returned by}
}
\value{
an MRG model
}
\description{
Load a mixed-regime Gaussian model from a fit vector
}
