% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{PCMNextMapping}
\alias{PCMNextMapping}
\title{Generate the next mapping of a set of model-types to a number of regimes}
\usage{
PCMNextMapping(mapping, models)
}
\arguments{
\item{mapping}{a vector with elements from \code{models}.
(repetitions are allowed) denoting a current model-to-regime mapping.}

\item{models}{a vector of PCM model-types or integer indices.}
}
\value{
a vector of the same length as mapping with elements from \code{models}.
}
\description{
Generate the next mapping of a set of model-types to a number of regimes
}
