% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{RetrieveBestFitScore}
\alias{RetrieveBestFitScore}
\title{Retrieve the best fit from a PCMFitMappings object.}
\usage{
RetrieveBestFitScore(fitMappings, rank = 1)
}
\arguments{
\item{fitMappings}{an object of S3 class 'PCMFitModelMappings' as returned by
\code{\link{PCMFitMixed}}.}

\item{rank}{an integer. Default: 1.}
}
\value{
a named list.
}
\description{
Retrieve the best fit from a PCMFitMappings object.
}
