% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{RetrieveBestModel}
\alias{RetrieveBestModel}
\title{Retrieve the optimal PCM model from a PCMFit object}
\usage{
RetrieveBestModel(fit)
}
\arguments{
\item{fit}{an object of S3 class \code{'PCMFit'}.}
}
\value{
a PCM model object.
}
\description{
Retrieve the optimal PCM model from a PCMFit object
}
\seealso{
\code{\link{PCMFit}}.
}
