% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{RetrieveFittedModelsFromFitVectors}
\alias{RetrieveFittedModelsFromFitVectors}
\title{Retrieve the ML fits from the fitVectors column in a table of fits.}
\usage{
RetrieveFittedModelsFromFitVectors(fitMappings,
  tableFits = fitMappings$tableFits,
  modelTypes = fitMappings$arguments$modelTypes, modelTypesNew = NULL,
  argsMixedGaussian = fitMappings$arguments$argsMixedGaussian,
  X = fitMappings$X, tree = fitMappings$tree, SE = fitMappings$SE,
  setAttributes = FALSE)
}
\arguments{
\item{fitMappings}{an object of S3 class PCMFitModelMappings.}

\item{tableFits}{a data.table}

\item{modelTypesNew}{NULL or a character vector containing all model-types in
fitMappings$arguments$modelTypes and, eventually, additional model-types.}

\item{setAttributes}{logical indicating if an X and tree attribute should be
set to each model-object. This is used for later evaluation of the
log-likelihood of the score for the model on the given tree and
data. Using a global tree for that is a bad idea, because the model may be
fit for a subtree, i.e. clade. Default FALSE.}
}
\value{
a copy of tableFits with added column "model" and, if necessary,
updated integer model-type indices in the "fitVector" column.
}
\description{
Retrieve the ML fits from the fitVectors column in a table of fits.
}
