% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{runOptim}
\alias{runOptim}
\title{Calling optim a number of times}
\usage{
runOptim(lik, parLower, parUpper, matParInit, control = NULL,
  verbose = TRUE)
}
\arguments{
\item{lik}{a function of numeric vector argument. Possible signatures are
function(par) {} or function(par, input) {}. This argument is mandatory and
doesn't have a  default value.}

\item{parLower, parUpper}{numeric vectors of equal length p: the number of
parameters of the function lik;}

\item{matParInit}{a numeric matrix of p columns.}

\item{control}{a list passed to optim().}

\item{verbose}{logical indicating whether informative messages should be
printed on the console during the run.}
}
\value{
a named list
}
\description{
Calling optim a number of times
}
\seealso{
\code{\link{configOptim}}.
}
