% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{UpdateCladeFitsUsingSubModels}
\alias{UpdateCladeFitsUsingSubModels}
\title{Update with parameters from submodels where the fit of a submodel is better}
\usage{
UpdateCladeFitsUsingSubModels(cladeFits, modelTypes, subModels,
  argsMixedGaussian, metaIFun = PCMInfo, scoreFun, X, tree, SE,
  verbose = FALSE)
}
\description{
Update with parameters from submodels where the fit of a submodel is better
}
