% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ComposeMixedGaussianFromFits}
\alias{ComposeMixedGaussianFromFits}
\title{Compose a MixedGaussian model}
\usage{
ComposeMixedGaussianFromFits(tree, startingNodesRegimes, modelTypes, k, R,
  mapping, argsMixedGaussian, tableFits, modelTypesInTableFits,
  tableAncestors = NULL, verbose = FALSE)
}
\description{
Compose a MixedGaussian model
}
