% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{LookupFit}
\alias{LookupFit}
\title{Lookup a fit vector for a given tree and model mapping in a data.table of
previously run fits.}
\usage{
LookupFit(tree, modelTypes, modelMapping, tableFits,
  hashCodes = HashCodes(tree = tree, modelTypes = modelTypes,
  startingNodesRegimesLabels =
  PCMTreeGetLabels(tree)[PCMTreeGetStartingNodesRegimes(tree)], modelMapping =
  modelMapping))
}
\arguments{
\item{tree}{a phylo object}

\item{modelTypes}{character vector}

\item{modelMapping}{an integer or character vector to by matched against modelTypes}

\item{tableFits}{a data.table having at least the following columns:
\itemize{
\item{hashCodeTree}{an MD5 key column of type character-vector}
\item{hashCodeStartingNodesRegimesLabels}{an MD5 key column of type character-vector representing the hash-code of
\code{PCMTreeGetLabels(tree)[PCMTreeGetStartingNodesRegimes(tree)]}.}
\item{hashCodeMapping}{an MD5 key column of type character-vector}}}
}
\value{
the corresponding fit-vector to the given tree and model mapping or
if no such entry is found, issues an error.
}
\description{
Lookup a fit vector for a given tree and model mapping in a data.table of
previously run fits.
}
