% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{MatchModelMapping}
\alias{MatchModelMapping}
\title{Match a model mapping vector against a vector of model types}
\usage{
MatchModelMapping(modelMapping, modelTypes)
}
\arguments{
\item{modelMapping}{a character or integer vector}

\item{modelTypes}{a character vector with valid model-class-names.}
}
\value{
a character vector with elements from modelTypes or stops with an error
}
\description{
Match a model mapping vector against a vector of model types
}
