% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{PCMFit}
\alias{PCMFit}
\title{Fitting a PCM model to a given tree and data}
\usage{
PCMFit(X, tree, model, SE = matrix(0, PCMNumTraits(model),
  PCMTreeNumTips(tree)), metaI = PCMInfo(X, tree, model, SE),
  positiveValueGuard = Inf, argsPCMParamLowerLimit = NULL,
  argsPCMParamUpperLimit = NULL, matParInit = NULL,
  numRunifInitVecParams = if (is.null(matParInit)) 1000L else 0L,
  numGuessInitVecParams = if (is.null(matParInit)) 100L else 0L,
  numCallsOptim = 10L, control = NULL, verbose = FALSE)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} and \code{NaN} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Missing values can be either not-available (\code{NA}) or not existing (\code{NaN}).
Thse two values have are treated differently when calculating
likelihoods: see \code{\link{PCMPresentCoordinates}}.}

\item{tree}{a phylo object with N tips.}

\item{model}{an S3 object specifying both, the model type (class, e.g. "OU") as
well as the concrete model parameter values at which the likelihood is to be
calculated (see also Details).}

\item{SE}{a k x N matrix specifying the standard error for each measurement in
X. Alternatively, a k x k x N cube specifying an upper triangular k x k
Choleski factor of the variance covariance matrix for the measurement error
for each node i=1, ..., N.
Default: \code{matrix(0.0, PCMNumTraits(model), PCMTreeNumTips(tree))}.}

\item{metaI}{a list returned from a call to \code{PCMInfo(X, tree, model, SE)},
containing meta-data such as N, M and k. Alternatively, this can be a
function object that returns such a list, e.g. the function\code{PCMInfo}
or the function \code{PCMInfoCpp} from the \code{PCMBaseCpp} package.}

\item{positiveValueGuard}{a real number (not necessarily positive) used during
the fit as a threshold for highly positive but incorrec log-likelihood values,
caused by numerical errors. In particular, such kind of errors cuased by
limited numerical precision have been observed when using the \code{PCMBaseCpp}
R-package (see argument \code{metaI} above). This argument value is \code{Inf}
by default and should be use only as a last escape. The recommended way to
of preventing such numerical errors is by setting more stringent values for
the runtime options \code{PCMBase.Threshold.EV} and
\code{PCMBase.Threshold.SV} (see \code{\link{PCMOptions}}).}

\item{argsPCMParamLowerLimit, argsPCMParamUpperLimit}{named lists with
arguments passed to the functions \code{\link{PCMParamLowerLimit}} and
\code{\link{PCMParamUpperLimit}}, respectively. Default: \code{NULL}.}

\item{verbose}{logical indicating if some debug-messages should printed.}
}
\value{
an object of class PCMFit
}
\description{
Fitting a PCM model to a given tree and data
}
\seealso{
\code{\link{PCMOptions}}
}
