% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitMixed.R
\name{PCMFitMixed}
\alias{PCMFitMixed}
\title{Fit regime-assignments to (sub-)trees in a tree with different assigned model
types to each regime.}
\usage{
PCMFitMixed(X, tree, modelTypes = MGPMDefaultModelTypes(),
  argsMixedGaussian = Args_MixedGaussian_MGPMDefaultModelTypes(),
  SE = matrix(0, nrow(X), PCMTreeNumTips(tree)),
  generatePCMModelsFun = GeneratePCMModelTypes, metaIFun = PCMInfo,
  positiveValueGuard = Inf, fitMappingsPrev = NULL,
  tableFitsPrev = fitMappingsPrev$tableFits,
  modelTypesInTableFitsPrev = NULL, listPartitions = NULL,
  minCladeSizes = 20L, maxCladePartitionLevel = if
  (is.null(listPartitions)) 8L else 1L,
  maxNumNodesPerCladePartition = 1L,
  listAllowedModelTypesIndices = c("best-clade-2", "best-clade", "all"),
  scoreFun = AIC,
  argsConfigOptim1 = DefaultArgsConfigOptim(numCallsOptim = 10),
  argsConfigOptim2 = DefaultArgsConfigOptim(numCallsOptim = 4),
  listPCMOptions = PCMOptions(), skipFitWhenFoundInTableFits = TRUE,
  doParallel = FALSE, prefixFiles = "fits_",
  saveTempWorkerResults = TRUE, printFitVectorsToConsole = FALSE,
  setAttributes = TRUE, verbose = TRUE, debug = FALSE)
}
\value{
an S3 object of class PCMFitModelMappings.
}
\description{
This function performs multiple model fits of mixed regime models
(MixedGaussian) mapping different model-types (e.g. BM and OU) to different
regimes (colors) in a tree and testing different regime assignments to the
branches in the tree.
}
