% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFit.R
\name{PCMFitModelMappings}
\alias{PCMFitModelMappings}
\title{Fit all possible model-type mappings to the regimes in a tree.}
\usage{
PCMFitModelMappings(X, tree, modelTypes = c("BM3", "OU3"),
  metaIFun = PCMInfo, lik = NULL, prior = NULL, input.data = NULL,
  config = NULL, fitToClades = FALSE,
  cladeRootNodes = PCMTreeGetStartingNodesRegimes(tree), fitToCladeRoot = 0,
  cladeFits = NULL, numJitterRootCladeFit = 100,
  sdJitterRootCladeFit = 0.5, numJitterAllCladeFits = 100,
  sdJitterAllCladeFits = 0.5, argsMRG = NULL, argsPCMLowerBound = NULL,
  argsPCMUpperBound = NULL, argsPCMSetOrGetVecParams = NULL,
  argsConfigOptimAndMCMC = NULL, printFitVectorsToConsole = FALSE,
  setAttributes = TRUE, doParallel = FALSE, verbose = FALSE)
}
\arguments{
\item{X}{a \code{k x N} numerical matrix with possible \code{NA} entries. Each
column of X contains the measured trait values for one species (tip in tree).
Depending on the value of the global option "PCMBase.Internal.PC.Full" \code{NA}
entries are interpreted either as missing measurements or non-existing traits
(see \code{\link{PCMPresentCoordinates}}).}

\item{tree}{a phylo object with N tips.}

\item{verbose}{logical indicating if some debug-messages should printed.}

\item{model}{a PCM model used primarily to extract attributes such as the
number of traits (k) and the set of model-types to use for the mappings.}
}
\value{
an object of class PCMSelect
}
\description{
This function performs multiple model fits of mixed regime models
(MRG) mapping different model-types (e.g. BM and OU) to the regimes in a tree.
The produced list of MRG-fits can be used for model selection based on information
criterions such as AIC.
}
