% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{PCMFitModelMappings}
\alias{PCMFitModelMappings}
\title{Fit regime-assignments to (sub-)trees in a tree with different assigned model
types to each regime.}
\usage{
PCMFitModelMappings(X, tree, modelTypes, SE = matrix(0, nrow(X),
  PCMTreeNumTips(tree)), generatePCMModelsFun = NULL,
  metaIFun = PCMInfo, positiveValueGuard = Inf, lik = NULL,
  prior = NULL, input.data = NULL, config = NULL,
  fitMappingsPrev = NULL, tableFitsPrev = fitMappingsPrev$tableFits,
  modelTypesInTableFitsPrev = NULL, skipFitWhenFoundInTableFits = TRUE,
  prefixFiles = "fits_", maxCladePartitionLevel = 8,
  maxNumNodesPerCladePartition = 1, minCladeSizes = 25,
  listPCMOptions = PCMOptions(), argsMixedGaussian = NULL,
  argsPCMParamLowerLimit = NULL, argsPCMParamUpperLimit = NULL,
  argsPCMParamLoadOrStore = NULL, argsConfigOptimAndMCMC1 = NULL,
  argsConfigOptimAndMCMC2 = NULL, numJitterRootRegimeFit = 100,
  sdJitterRootRegimeFit = 0.5, numJitterAllRegimeFits = 100,
  sdJitterAllRegimeFits = 0.5, printFitVectorsToConsole = FALSE,
  setAttributes = TRUE, doParallel = FALSE, verbose = TRUE,
  verbosePCMFit = FALSE, verboseComposeMixedGaussianFromFits = FALSE,
  verboseAdaptArgsConfigOptimAndMCMC = FALSE)
}
\value{
an S3 object of class PCMFitModelMappings.
}
\description{
This function performs multiple model fits of mixed regime models
(MixedGaussian) mapping different model-types (e.g. BM and OU) to different
regimes in a tree and testing different regime assignments to the branches in
the tree.
}
