% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCMFitModelMappings.R
\name{PCMLoadMRGFromFitVector}
\alias{PCMLoadMRGFromFitVector}
\title{Load a mixed-regime Gaussian model from a fit vector}
\usage{
PCMLoadMRGFromFitVector(fitVector, modelTypes, k, ...)
}
\description{
Load a mixed-regime Gaussian model from a fit vector
}
