% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{PCMNextMapping}
\alias{PCMNextMapping}
\title{Generate the next mapping of a set of models to a number of regimes}
\usage{
PCMNextMapping(mapping, models)
}
\arguments{
\item{mapping}{a character vector with elements from models
(repetitions are allowed) denoting a current model-to-regime mapping.}

\item{models}{a character vector with the class-names of PCM models}
}
\value{
a character vector of the same length as mapping with elements from models.
}
\description{
Generate the next mapping of a set of models to a number of regimes
}
